// src/types/data/report/abstruct-budget-type.ts
export interface GetSingleAbstructBudgetItemShape {
  id: number;
  areaName: string;
  mosavabRevenue: number;
  mosavabCurrent: number;
  mosavabCivil: number;
  mosavabFinancial: number;
  mosavabSanavati: number;
  mosavabPayMotomarkez: number;
  mosavabDar_Khazane: number;
  balanceMosavab: number;
  resoures: number;
}